Food Forecasting on Fulfillment Centers
I have worked on a Food forecasting Tableau project, where I had to help the client understand their meal delivery fulfillment centers' performance and stock up the raw materials as per the demand. Using Tableau, I developed a dashboard and story for the higher-ups to understand the granularity of my work in this project. It talked about the level of demand, the cost of items, and the average operational areas of the meal fulfillment centers. The region-wise, and cuisine-wise data was visualized and structured together, in the story.